requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
colorama>=0.4.6 
psutil
mutagen>=1.47.0
PyNaCl>=1.5.0 